restructure content-model
this changes the content-model for talks and pages. Mainly, the
association of the two is reversed, so every speaker will have a
talk-reference instead of the talk having a speaker-reference. This
better reflects the fact that there might be multiple speakers per talk
but - at least at our conference - not multiple talks per speaker.

The other change is to use the content of the markdown files for the
main content (speaker copy-text and talk-description).
This allows us to use all sorts of markup in there and should make it a
bit easier to edit.